feat:增加了Jenkinsfile脚本文件

+++ src/main/java/com/lzh/lol/po/Bet.java
@@ -37,7 +37,7 @@ public class Bet {
 
     private Integer betTeamId;
 
-    private Integer betScore;
+    private Double betScore;
 
     private Date createDate;
 }
