+++ src/main/java/com/lzh/lol/vo/BetVO.java
@@ -23,6 +23,6 @@ public class BetVO {
 
     private Integer betTeamId;
 
-    private Integer betScore;
+    private Double betScore;
 
 }
